+++ src/Kean.Core.Serialize/Serializer/SignedByte.cs
@@ -38,7 +38,7 @@ namespace Kean.Core.Serialize.Serializer
 		}
 		public object Deserialize(Storage storage, Reflect.Type type, Data.Node data)
 		{
-			return data is Data.SignedByte ? (data as Data.SignedByte).Value : 0;
+			return data is Data.SignedByte ? (data as Data.SignedByte).Value : (sbyte)0;
 		}
 		#endregion
 	}
